Remove switchComparison
